Move TS output
This makes browsing the `build` directory much easier as it is not
cluttered with `.js` and `.js.map` files.

The downside is that `resolveJsonModule` has to be disabled which
removes the automatic typing of `package.json`. That option has to be
disabled in order for `rootDir` to be defined since `package.json` is
outside of `rootDir`. `rootDir` has to be defined so that `rootPath`
gets determined correctly in the compiled JS.